antialiasing

+++ src/ray.hpp
@@ -20,7 +20,7 @@ class Ray {
     public:
         Ray() : _origin(vec3(0,0,0)), _direction(vec3(0,1,0)), _color(vec3(255,255,255)), _depth(4) {}
 
-        Ray(int x, int y, int width, int height, float fov, int depth = 4) { //Primary ray
+        Ray(float x, float y, int width, int height, float fov, int depth = 4) { //Primary ray
             float aspectRatio = width / (float)height; // Si width > height
             float x_camera = (2* ((x + 0.5) / width)- 1) * aspectRatio * tan(fov/2 * M_PI /180); 
             float y_camera = (1 - 2*((y + 0.5) / height)) * tan(fov/2 * M_PI/180);
